requirements.txt
streamlit
pandas
numpy
scikit-learn==1.7.2
psycopg2-binary